// src/types/product.ts
export type Product = {
	id: number;
	title: string;
	description: string;
	link: string;
	image_link: string;
	additional_image_link: string;
	availability: "In stock" | "Out of stock" | "Pre-order";
	list_price: string;
	sale_price: string;
	gtin: string;
	product_type: string;
	brand: string;
	condition: "New" | "Used" | "Refurbished";
	raw_color: string;
	color: string;
	gender: "female" | "male" | "unisex";
	size_format: "INT" | "IT" | "FR" | "Belts cm" | "ONE-SIZE";
	sizing_schema: string;
	sizes: string;
	size_type: "Regular" | "Plus" | "Petite";
	item_group_id: number;
	category: "Clothing" | "Accessories" | "Bags";
	shipping: string;
	mpn: string;
	material: string;
	collection: string;
	additional_image_link_2: string;
	additional_image_link_3: string;
	additional_image_link_4: string;
};